\name{Perm.CI2}
\alias{Perm.CI2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Permutation test inference for a treatment effect on a
  binary outcome}
\description{Computes the permutation confidence set for the average
  treatment effect on a binary outcome; compiled by \pkg{compiler} and used as Perm.CI}

\seealso{
\code{\link{Perm.CI}}
}