\name{pval2}
\alias{pval2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Computes permutation test p-value  
}
\description{Returns permutation test p-value; used in the construction of permutation test confidence sets by the function Perm.CI; compiled by \pkg{compiler} and used as pval
}

\seealso{
\code{\link{pval}}
}