\name{RI2by2-package}
\alias{RI2by2-package}
\alias{RI2by2}
\docType{package}
\title{Randomization inference for treatment effects on a binary outcome}
\description{Computes attributable effects based confidence set, 
  permutation test confidence set, and asymptotic confidence interval
  for the average treatment effect on a binary outcome in experiments where \eqn{m} of
  \eqn{n} subjects are randomized to treatment.}
\details{
\tabular{ll}{
Package: \tab RI2by2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-02\cr
License: \tab GPL-3\cr
}
}

\author{Joseph Rigdon \email{jrigdon@bios.unc.edu}
}
\references{Rigdon, J. and Hudgens, M.G. ``Randomization inference for treatment effects on a binary outcome.'' Submitted to Statistics in Medicine 2014.}


